Update patch set 3
Patch Set 3:

This is the other half of the original frameworks/base change that was previously +2'd, merged into AOSP master, and then lost in the 4.3 merge. Needs to be taken with the other change; both have the same Change-Id.

Patch-set: 3
Reviewer: Stephen Smalley <1010111@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0