Update patch set 3
Patch Set 3:

Obviously we can't resolve that, so can someone resolve it internally?  Or do we have to wait yet another release cycle to see it in AOSP?  I think these patches have been in gerrit for over a year now (in the original form, before installd was moved to frameworks/native).

Patch-set: 3