Update patch set 3
Patch Set 3:

Where do we stand on this change?
Is there still an intent to support policy updates via SELinuxPolicyInstallReceiver or some other mechanism other than an OTA?  If so, then we still need this to refresh installd's seapp_contexts upon a policy update.  If not, then I am not clear that we need this anymore.

Patch-set: 3
Reviewer: Stephen Smalley <1010111@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0