Update patch set 1
Patch Set 1:

I checked with the host GL Tracer, and as far as I see, the only pointer value it tries to read is the pointer argument that is part of glVertexAttribPointer(). This pointer is saved and displayed as part of the vertex attribute state.

I'll submit a separate CL fixing the host tool. Once this change goes in, users would require the updated host tool for things to work properly.

Patch-set: 1