Update patch set 2
Patch Set 2:

I'm surprised there is a bug here -- but I can totally believe it. I think the best thing to do, would be to write a small  test that checks that Transform.cpp agrees with this change (it calculates the transform bits from the 2x2 matrix). It might even make sense to have a static method on Transform.cpp that does this inversion, so it's all in one place.

Patch-set: 2
Reviewer: Mathias Agopian <1021189@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0