Update patch set 2
Patch Set 2:

The goal is to split up parts of /data/misc/keychain so that each user has their own version of security-related things, rather than using whatever user 0 has set. For example currently if user 0 trusts DODGY_CA.crt as a certificate authority so does everybody else because they all look in the same directory for CAs.

It's difficult to make this work via an API because most of the files in keychain are used by conscrypt. Since conscrypt is tied to the Security architecture and not dependent on the platform it can't talk directly to system services.

There are ways of getting the same effect by defining a ContentProvider-style interface in conscrypt and then having the system provide an implementation, but that seems very hacky in comparison.

Patch-set: 2
Label: Verified=0