Update patch set 2
Patch Set 2: Code-Review+1

the flags argument isn't a general purpose solution due to the limit on length (92 chars?) for the value. fixing this is fine for debugging however.

I plan to eventually split the ones we actually need such as compiler-filter to dedicated properties. it also allows more precise overriding as well. for example, i need to fix AndroidRuntime.cpp to override filter when we are booting with encrypted file system, hard to do that with flags overriding that potentially.

Patch-set: 2
Reviewer: Brian Carlstrom <1003723@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1