Update patch set 1
Patch Set 1:

The old code certainly depends on undefined behavior and best to be changed.
But that would probably involve many callers. It's very possible that the code will work with Clang's implementation if the warning is suppressed.
Casting 'this' to void* should confuse the Clang compiler so it won't be able to detect, give warning, or optimize out the compare-to-null logic.

Patch-set: 1