Update patch set 1
Patch Set 1:

Undefined behavior is precisely why it warns in this case. You cannot guarantee what the compiler does, and they can (and will) break your code. GCC and LLVM are both moving very aggressively in this area. You do not want to leave this as-is. If we are actually relying on UB, we need to fix the root cause.

Patch-set: 1