Update patch set 1
Patch Set 1: -Code-Review

Mmm, okay.  I read that as "it's undefined behavior, and in our compiler we are choosing to define the behavior [this way]" -- which is unfortunately the definition that makes it hardest to defend against malformed code we don't control.

So I guess we have no choice but to crash rather than fail in-band, because the compiler is going to enforce that behavior anyway?  Well then.

Patch-set: 1
Reviewer: Chris Tate <1000322@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0