Update patch set 1
Patch Set 1: Code-Review-1

> Mmm, okay.  I read that as "it's undefined behavior, and in our
 > compiler we are choosing to define the behavior [this way]" --
 > which is unfortunately the definition that makes it hardest to
 > defend against malformed code we don't control.

That is "unspecified behavior", and it is treated differently from undefined behavior. We are trying to minimize the differences there (for LLVM and gcc). The real problem is that either or both of these compilers can elide the check completely on a compiler upgrade, and still be completely correct.

 > 
 > So I guess we have no choice but to crash rather than fail in-band,
 > because the compiler is going to enforce that behavior anyway? 
 > Well then.

As Dan just mentioned, we have to fix all of the callers here first, since they are indeed relying on UB to solve this problem for them. Once we fix the callers, we can submit a patch to remove the comparisons.

Patch-set: 1
Label: Code-Review=-1