Update patch set 1
Patch Set 1:

> What is libsigchain and what is the effect of linking it in
 > surfaceflinger?

libsigchain was previously an LD_PRELOAD in init.rc for proper libart.so behavior but nnk wants us to move away from that.


 > SurfaceFlinger is entirely independent of ART

not actually. There is this mysterious debug feature when debug.sf.ddms is set:

frameworks/native/services/surfaceflinger/DdmConnection.cpp
    // TODO: Should this just link against libnativehelper and use its
    // JNI_CreateJavaVM wrapper that essential does this dlopen/dlsym
    // work based on the current system default runtime?
    void* libart_dso = dlopen("libart.so", RTLD_NOW);

if this is obsolete and can be removed, we can abandon all of this effort.

Patch-set: 1
Reviewer: Brian Carlstrom <1003723@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0