Update patch set 3
Patch Set 3:

> Any news on this? Internal master is still crashing because e.g.
 > media code will often try to put a NULL Binder in a parcel, and the
 > asBinder call will segfault.

If you are putting NULL into a parcel, you need to check for NULL before calling asBinder(). Please update any code you see that isn't doing this. If you don't, it will continue to remain broken, as we are not going to play a game of chicken with C++ undefined behavior.

Patch-set: 3