Update patch set 1
Patch Set 1:

> I've wanted to this for a while as well.
 > Last time I looked at it, it was more than just bring-up that
 > needed it. Recovery

not any more. see bootable/recovery/minui.

(speaking of which, if anyone with a clue about graphics would like to help me debug why text rendering is slow and whether we can do anything about it...)

an OEM (i forget which) recently asked for minui as a static library for use in their factory test images. is that something that could be useful for bringup work? do we know what their specific constraints are?

 > (and possibly the first boot on encrypted
 > devices -- not sure) used pixelflinger for graphics.

i don't think so. libGLES_android seems to be the only pixelflinger user, and that's only used here:

https://cs.corp.google.com/#search/&q=libGLES_android%20file:%5C.mk&sq=package:%5Eandroid$&type=cs&m=100&sort=1

 > Is there a way we could add restrictions in the build-system, such
 > that new dependencies couldn't be added?

that would certainly be useful in general when we're trying to clean up the tree.

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0