Update patch set 1
Patch Set 1:

> an OEM (i forget which) recently asked for minui as a static
 > library for use in their factory test images. is that something
 > that could be useful for bringup work? do we know what their
 > specific constraints are?

At some point, libagl was "good enough" to run basic apps, so you could actually boot the real system. I think that's what people doing bring-up are after. 

In fact we had to keep a ES1.x code path in SF when we switched to ES2.x just for that purpose -- there are no real h/w using ES1.x anymore.

it's going to be harder and harder to keep it running and in fact it's hurting the system (memory, bugs, code complexity, maintenance burden)

A "solution" that I was pushing for, was to use Mesa for bring-up -- I don't know what happened with that. It would just present itself as a gl driver.

Patch-set: 1
Reviewer: Mathias Agopian <1001056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0