Update patch set 1
Patch Set 1:

The behavior as it is implemented today is by design. If a producer queues some buffers, disconnects, connects, and queues some more buffers, we don't want to lose the buffers in the queue at the time of disconnect. Are you seeing a case where this is causing problems?

Patch-set: 1