Update patch set 1
Patch Set 1:

I understand that you're mostly concerned about M but don't you think something like that should be fixed in Lollipop as well?

Adding another flag is not the most elegant solution but it's simple and it works in Lollipop without breaking existing apps. From the pragmatic point of view that should be enough to merge it even if it's going to be reverted in M.

If you think it's not worth it, could you propose some alternative fix for the Lollipop MediaProjection tapjacking issue? https://android-review.googlesource.com/#/c/157584/

Patch-set: 1