Update patch set 1
Patch Set 1:

The output is relatively short and compressible, so it is not a big cost. I consider this addition a 2016 goal.

As I am not the spam police, just replace 'I' with any person wishing to be one below.

Every time someone declares that something is spammy, or that they lost their logs because they were spammy (or related to other threads in the same UID pool that is spammy) I have to ask them to report 'logcat -b all -S'. Every time I want to spawn a new 'stop being so spammy' bug spawned off a bugreport collected for another purpose, I *wish* I had the 'logcat -b all -S' report.

Patch-set: 1