Update patch set 1
Patch Set 1: Verified+1

NB: to be clear, I mean the visible consumer of the value. I have, of course, unit tested this all with LOG messages to confirm the ashmem_get_size_region() calls are getting through the gate and still being made for _many_ of the incoming file descriptors, my concern is that _all_ of the calls are getting through, and that the math of + and - adds up.

Patch-set: 1
Label: Verified=+1