Update patch set 5
Patch Set 5:

Hi Dan,

I've just encountered this problem locally on an older android L tree. I wrote a similar patch and was about to upstream it, when I realised you've independently found and fixed the same issue!

One thing this patch doesn't address is when IPC is between a 32bit and 64bit process. I think it will break here, as some member variables of the BufferItem are not explicitly sized. For example, the members that use 'int' and  'bool'. These may be written/read with different sizes/alignment on a 32bit process vs 64bit.

My patch did try to address this, by updating the logic in flatten/unflatten to basically write 'int' and 'bool' as int32_t's, and convert to/from them. My patch no longer applies on android M, as this code has changed quite a bit since android L. Even so, it probably wasn't the nicest solution anyway.

I think the correct solution is to update the types used on the BufferItem class header, so that it doesn't use 'bool' or 'int', but instead int32_T or equivalent.

Patch-set: 5
Reviewer: Francis Hart <1044843@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0