Update patch set 1
Patch Set 1: Code-Review-1

Is this in response to an issue you've actually seen? Typically a method called *Locked is assumed to be called with the lock held, which would mean that it would deadlock if you tried to grab the lock again. As far as I can tell, this method is only called from ConsumerBase::abandon, which grabs the lock before calling into either ConsumerBase::abandonLocked or GLConsumer::abandonLocked (which calls back into ConsumerBase::abandonLocked) with the lock held.

Patch-set: 1
Label: Code-Review=-1