Update patch set 1
Patch Set 1:

I think the confusion comes from whether status_t is a generic or a binder-specific error reporting mechanism.  I think it is generic in that

 - utils/Errors.h has no mention of libbinder
 - there are error codes that are never reported by libbinder

When status_t is generic, it could be used as the error code for EX_SERVICE_SPECIFIC.  Also, it deserves to be a first-class error and be supported from binder::Status.

Patch-set: 1