Update patch set 1
Patch Set 1:

You aren't generating ServiceSpecificExceptions today, unless you've gone and intentionally written code to do so.  This is unlikely, since service specific exceptions are a new concept we introduced with aidl-cpp.  What you think you are doing is using the mechanism that Java uses to report transport errors to report service specific exceptions.  What you're actually doing in that code is to return NO_ERROR, and writing a success value to the parcel.  It would correspond to the following code:

That code you presented seems to correspond to:

interface IGraphicsBufferProducer {
  int SetAutoRefresh(int value);
}

and you would always return Status::ok() in this case.  The returned value is actually how you're signalling error.

Patch-set: 1